test(py#agent,py#fast-api): split heavy spec files for parallelism
vitest runs each spec file serially in a single worker, so the largest
file caps the whole suite's wall-clock regardless of sharding.
py/agent/generator.spec.ts (583s) and py/fast-api/generator.spec.ts
(248s) were by far the largest, flooring shard times at ~10 min.

Split them into balanced sibling files, preserving the top-level
describe block names so snapshot keys are unchanged. Snapshot entries are
moved verbatim into matching per-file .snap files:
- py/agent -> core / constructs / protocols / frameworks (max ~152s)
- py/fast-api -> main + terraform (max ~147s)

No test logic or snapshot content changed.
